refactor: 프로필 페이지에서 사용하는 쿼리키 정리
