+++ src/queries/bookshelf/useMySummaryBookShelfQuery.ts
@@ -2,10 +2,11 @@ import bookshelfAPI from '@/apis/bookshelf';
 import { useQuery } from '@tanstack/react-query';
 import type { QueryOptions } from '@/types/query';
 import type { APIBookshelf } from '@/types/bookshelf';
+import bookShelfKeys from './key';
 
 const useMySummaryBookshlefQuery = (options?: QueryOptions<APIBookshelf>) =>
   useQuery(
-    ['summaryBookshlef', 'me'],
+    bookShelfKeys.summary('me'),
     () => bookshelfAPI.getMySummaryBookshelf().then(({ data }) => data),
     options
   );
